64bit libao2 fix by Jens Axboe <mplayer-dev@kernel.dk>
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@9634 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libao2/audio_out.c
@@ -124,7 +124,7 @@ ao_functions_t* init_best_audio_out(char** ao_list,int use_plugin,int rate,int c
 	    if(!strcmp(audio_out->info->short_name,ao)){
 		// name matches, try it
 		if(use_plugin){
-		    audio_out_plugin.control(AOCONTROL_SET_PLUGIN_DRIVER,(int)audio_out);
+		    audio_out_plugin.control(AOCONTROL_SET_PLUGIN_DRIVER,audio_out);
 		    audio_out=&audio_out_plugin;
 		}
 		if(audio_out->init(rate,channels,format,flags))
@@ -140,7 +140,7 @@ ao_functions_t* init_best_audio_out(char** ao_list,int use_plugin,int rate,int c
     for(i=0;audio_out_drivers[i];i++){
 	ao_functions_t* audio_out=audio_out_drivers[i];
 	if(use_plugin){
-	    audio_out_plugin.control(AOCONTROL_SET_PLUGIN_DRIVER,(int)audio_out);
+	    audio_out_plugin.control(AOCONTROL_SET_PLUGIN_DRIVER,audio_out);
 	    audio_out=&audio_out_plugin;
 	}
 	if(audio_out->init(rate,channels,format,flags))
